Partial solution for Day 19 challenges (3)
During the reactor I decided to re-think the whole challenge.
I attempted to prepare a series of tests and assertions to verify
my assumptions and understanding the whole task. During this
I understood what was the main problem with nested rules
– one such can be matched multiple times (various ways) to the same
input string. For example, rules like `5: "b"` and `1: 5 | 5 1` can
match `b`, `bb`, `bbb` and so on. Now it is important to notice
that not always the longest match would be the correct one;
considering the rule like `0: 1 1`, it should match properly
the input `bb` and `bbb` (and it would not in naive "always longest"
approach). This was mainly the case why my previous attempts failed
on the real input data.