Optimize solution for Day 12 / part 2
Browsing the internet, I found very similar solution,
just using the deque instead of list. I checked out of
curiosity and changing it in my code resulted in much
greater performance. Therefore I read about differences
between list and deque in Python and apparently the latter
was optimized for removing the first element.

As in the solution the order in which we process list elements
does not matter, I changed pop(0) to pop() in my original code
and it results in 4 times faster program. It is interesting
that in smaller problems (e.g. part 1) it is not that significant.
Still – a lesson learned.