Add script for testing times
The time itself is not crucial in Advent of Code,
but I thought that it would be nice to see if there
are some solutions that are slower than the other ones.
This script calculates an average of 10 runs and returns
colored output to let me know which solution may I need
to improve (provided that I have some free time...).