Optimize solution for Day 23 / part 2 once more
I noticed that the condition was not correct – from the right side
(index2) we should never subtract, while from the left side (index1)
we can always subtract, as this position never matters (it is either
a current element we want to omit or a position next to room which
is always empty). Interestingly this further reduces the computation
time on average from 3.2 second to 3.0 second.