Replace hashlib with own MD5 implementation
To remain compatible with the no-import challenge, I had
to implement the MD5 algorithm myself, so there can be
a solution without using the hashlib module. The prepared
code works, however it about two orders of magnitude slower
when executed with CPython3. However, when executed with PiPy,
the execution is only about one order of magnitude slower,
which is acceptable enough for these days.

Project: https://github.com/sdatko/MD5Py